Haciendo metodos de estado y poblacion. Cambiando estructura.

+++ src/interfaz/Cobrable.java
@@ -2,6 +2,6 @@ package interfaz;
 
 public interface Cobrable {
 
-	public void cobrar(int sueldo, double dineroEstado);
+	public void cobrar(double sueldo, double dineroEstado);
 
 }
